requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Database and vector operations
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# Document processing
pdfplumber>=0.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# LLM backends
llama-cpp-python>=0.2.0
openai>=1.0.0

# Configuration and utilities
python-dotenv>=1.0.0
rich>=13.0.0
tqdm>=4.65.0

# Optional ML dependencies
numpy>=1.24.0
scikit-learn>=1.3.0

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0